requirements.txt
APScheduler==3.10.1
py-cord==2.4.0
python-dotenv==1.0.0
requests==2.28.2